Data Science/requirements.txt
numpy==2.3.4
pandas==2.3.3
scipy==1.16.2
scikit-learn==1.7.2
plotly==6.3.1
openpyxl==3.1.5
fpdf2==2.8.4
nbformat==5.10.4
nbconvert==7.16.6
ipykernel==7.0.1
kaleido
